moved code out of __int__.py and into app.py
